Revert to a older version of `SnapshotChessBoardState`
