+++ mobile/src/main/java/ch/epfl/sdp/mobile/state/StatefulGameScreen.kt
@@ -10,7 +10,7 @@ import ch.epfl.sdp.mobile.application.chess.engine.*
 import ch.epfl.sdp.mobile.application.chess.engine.Color.Black
 import ch.epfl.sdp.mobile.application.chess.engine.Color.White
 import ch.epfl.sdp.mobile.application.chess.notation.serialize
-import ch.epfl.sdp.mobile.state.BasicSnapshotBoardState.SnapshotPiece
+import ch.epfl.sdp.mobile.state.SnapshotChessBoardState.SnapshotPiece
 import ch.epfl.sdp.mobile.ui.game.*
 import ch.epfl.sdp.mobile.ui.game.GameScreenState.Message
 import ch.epfl.sdp.mobile.ui.game.GameScreenState.Message.*
@@ -81,7 +81,7 @@ class SnapshotChessBoardState(
     private val user: AuthenticatedUser,
     private val match: Match,
     private val scope: CoroutineScope,
-) : GameScreenState<SnapshotPiece>, BasicSnapshotBoardState(match, scope) {
+) : GameScreenState<SnapshotPiece> {
 
   // TODO : Implement these things.
   override var listening by mutableStateOf(false)
@@ -94,6 +94,7 @@ class SnapshotChessBoardState(
 
   override fun onBackClick() = actions.onBack()
 
+  private var game by mutableStateOf(Game.create())
   private var whiteProfile by mutableStateOf<Profile?>(null)
   private var blackProfile by mutableStateOf<Profile?>(null)
 
@@ -118,13 +119,43 @@ class SnapshotChessBoardState(
   }
 
   init {
+    scope.launch { match.game.collect { game = it } }
     scope.launch { match.white.collect { whiteProfile = it } }
     scope.launch { match.black.collect { blackProfile = it } }
   }
 
+  /* */
+  /**
+   * An implementation of [ClassicChessBoardState.Piece] which uses a [PieceIdentifier] to
+   * disambiguate different pieces.
+   *
+   * @param id the unique [PieceIdentifier].
+   * @param color the color for the piece.
+   * @param rank the rank for the piece.
+   */
+  data class SnapshotPiece(
+      val id: PieceIdentifier,
+      override val color: ChessBoardState.Color,
+      override val rank: ChessBoardState.Rank,
+  ) : ChessBoardState.Piece
+
   /** The currently selected [Position] of the board. */
   override var selectedPosition by mutableStateOf<ChessBoardState.Position?>(null)
     private set
+
+  override val checkPosition: ChessBoardState.Position?
+    get() {
+      val nextStep = game.nextStep
+      if (nextStep !is NextStep.MovePiece || !nextStep.inCheck) return null
+      return game.board
+          .first { (_, piece) -> piece.color == nextStep.turn && piece.rank == Rank.King }
+          .first
+          .toPosition()
+    }
+
+  override val pieces: Map<ChessBoardState.Position, SnapshotPiece>
+    get() =
+        game.board.asSequence().map { (pos, piece) -> pos.toPosition() to piece.toPiece() }.toMap()
 
   override val availableMoves: Set<ChessBoardState.Position>
     // Display all the possible moves for all the pieces on the board.
@@ -187,4 +218,29 @@ class SnapshotChessBoardState(
 
   override val moves: List<Move>
     get() = game.serialize().map(::Move)
+}
+
+fun Piece<Color>.toPiece(): SnapshotPiece {
+  val rank =
+      when (this.rank) {
+        Rank.King -> ChessBoardState.Rank.King
+        Rank.Queen -> ChessBoardState.Rank.Queen
+        Rank.Rook -> ChessBoardState.Rank.Rook
+        Rank.Bishop -> ChessBoardState.Rank.Bishop
+        Rank.Knight -> ChessBoardState.Rank.Knight
+        Rank.Pawn -> ChessBoardState.Rank.Pawn
+      }
+
+  val color =
+      when (this.color) {
+        Black -> ChessBoardState.Color.Black
+        White -> ChessBoardState.Color.White
+      }
+
+  return SnapshotPiece(id = this.id, rank = rank, color = color)
+}
+
+/** Maps a game engine [Position] to a [ChessBoardState.Position] */
+fun Position.toPosition(): ChessBoardState.Position {
+  return ChessBoardState.Position(this.x, this.y)
 }
